backend-structure-hotfix - Created common RepositoryInterface. Repo interfaces moved into contracts folder.
