+++ app/QuestionService/QuestionService.php
@@ -2,12 +2,13 @@
 namespace App\QuestionService;
 use App\Exceptions\QuestionServiceException;
 use App\QuestionService\Contracts\QuestionServiceInterface;
-use App\Repositories\Repositories\QuestionRepository;
-use App\Repositories\Repositories\AnswerRepository;
+use App\Repositories\Contracts\QuestionRepository;
+use App\Repositories\Contracts\AnswerRepository;
 use App\Exceptions\RepositoryException;
 use Response;
 
-class QuestionService implements QuestionServiceInterface {
+class QuestionService implements QuestionServiceInterface
+{
 
     private $questionRepository;
     private $answerRepository;
@@ -21,7 +22,11 @@ class QuestionService implements QuestionServiceInterface {
     }
     
     public function createQuestion($data){}
-    
+
+    /**
+     * @param $id
+     * @return model
+     */
     public function getQuestion($id)
     {
         try {
@@ -29,7 +34,7 @@ class QuestionService implements QuestionServiceInterface {
                 ->findWithRelations($id, ['user', 'folder']);
         } catch (RepositoryException $e) {
             throw new QuestionServiceException(
-                $e->getMessage() . " No such question",
+                $e->getMessage() . ' No such question',
                 null,
                 $e
             );
@@ -37,7 +42,11 @@ class QuestionService implements QuestionServiceInterface {
 
         return $question;
     }
-    
+
+    /**
+     * @param $question_id
+     * @return Collection
+     */
     public function getAnswersOfQuestion($question_id)
     {
         return $this->answerRepository
