+++ app/Repositories/Repositories/FolderRepositoryEloquent.php
@@ -4,7 +4,7 @@ namespace App\Repositories\Repositories;
 
 use Prettus\Repository\Criteria\RequestCriteria;
 use App\Repositories\Entities\Folder;
-
+use App\Repositories\Contracts\FolderRepository;
 /**
  * Class FolderRepositoryEloquent
  * @package namespace App\RepositoriesRepositories;
