+++ app/Repositories/Repositories/Repository.php
@@ -5,8 +5,10 @@ namespace App\Repositories\Repositories;
 use Prettus\Repository\Eloquent\BaseRepository;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use App\Exceptions\RepositoryException;
+use App\Repositories\Contracts\RepositoryInterface;
 
-abstract class Repository extends BaseRepository {
+abstract class Repository extends BaseRepository implements RepositoryInterface
+{
 
     /**
      * Find data by id
@@ -24,11 +26,23 @@ abstract class Repository extends BaseRepository {
         }
     }
 
+    /**
+     * @param $id
+     * @param array $relations
+     * @return model
+     */
     public function findWithRelations($id, array $relations)
     {
         return $this->with($relations)->find($id);
     }
 
+    /**
+     * @param $fieldName
+     * @param $fieldValue
+     * @param $relations
+     * @param array $columns
+     * @return collection
+     */
     public function findByFieldWithRelations($fieldName, $fieldValue, $relations, $columns=['*'])
     {
         return $this->with($relations)->findByField($fieldName, $fieldValue);
