+++ database/seeds/FoldersSeeder.php
@@ -1,7 +1,7 @@
 <?php
 
 use Illuminate\Database\Seeder;
-use App\Repositories\Repositories\FolderRepository;
+use App\Repositories\Contracts\FolderRepository;
 use Faker\Factory;
 
 class FoldersSeeder extends Seeder
